dlt-project/pyproject.toml
[project]
name = "dlt_project"
requires-python = ">=3.9,<3.14"
version = "0.1.0"
dependencies = [
    "dagster==1.12.1",
    "dagster-dlt>=0.28.1",
    "dpath>=2.2.0",
    "google-cloud-bigquery>=3.38.0",
    "google-cloud-bigquery-storage>=2.34.0",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
force-include = { "pyproject.toml" = "pyproject.toml" }

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "dlt_project"
registry_modules = [
    "dlt_project.components.*",
]


